Rename time shim; avoid name clash with real time package.
